+++ NEWS.md
@@ -3,6 +3,12 @@ Phan NEWS
 ?? ??? 2018, Phan 1.0.4 (dev)
 -----------------------
 
+Language Server/Daemon mode:
++ Support hover text for variables.
+
+  - For union types with a single non-nullable class/interface type, the hover text include the full summary description of that class-like.
+  - For non-empty union types, this will just show the raw union type (e.g. `string|false`)
+
 07 Sep 2018, Phan 1.0.3
 -----------------------